Finalizaçao do codigo teogym

+++ app/Http/Controllers/Admin/ExercicioController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreUpdateExercicioRequest;
 use App\Models\Exercicio;
+use App\Models\TreinoExercicio;
 use Illuminate\Http\Request;
 
 class ExercicioController extends Controller
@@ -128,10 +129,32 @@ class ExercicioController extends Controller
             abort(403);
         }
         if(!$exercicio = $this->repository->find($id))
-            dd($exercicio);
+            return redirect()->back()->with('warning', "Ooops! este exercicio não existe");
+
+        $treinos = TreinoExercicio::where('exercicio_id', $id)->get();
+        if($treinos->count() > 0) {
+            return redirect()->back()->with('warning', "Ooops! este exercicio já está em uma rotina e não pode ser deletado");
+        }    
 
         $exercicio->delete();
         
         return redirect()->route('exercicios.index');
+    }
+
+    /**
+     * Search results
+     *
+     * @param  \Illuminate\Http\Request $request
+     * @return \Illuminate\Http\Response
+     */
+    public function search(Request $request) 
+    {
+        $filters = $request->except('_token');
+        $exercicios = $this->repository->search($request->filter);
+
+        return view('app.exercicios.index', [
+            'exercicios' => $exercicios,
+            'filters' => $filters
+        ]);
     }
 }
